limit the state updates to a maximum updateRate
Van still provides the tools to limit the update rate. 

```JS
let {updateRate} = van
updateRate(100) 
```

This set´s the update rate to 100ms. state changes are delayed by up to 100 ms, before they are propagated to the DOM. If multiple changes are applied withi this time, only the last is updating the DOM.